pkg: fileutil: take into account hard links
When calculating directory size.

Also, rearrange the linux/unsupported go files.
